Update simple-interest.sh
Updated the simple-interest.sh file to include my GitHub username in the authors section. This acknowledges my contribution to the project and allows for easy identification and contact. I'm excited to be part of this initiative to empower individuals through micro-finance. Please note that the script is for sample purposes only and caution should be exercised when using or modifying it.